WIP: refactor into classes, brushstroke class approach
