+++ index.html
@@ -13,11 +13,11 @@
 					<!-- <input type="color" id="color-picker" value="#ffaaff"> -->
 					<div class="category-toolbar"></div>
 				</div>
-				<div class="drawing-toolbar">
+				<div class="effect-toolbar">
 					<button id="clear-all-button"class="light-button drawing-control-button">clear all</button>
 					<button id="download-button" class="light-button drawing-control-button">download</button>
-					<button id="redo-button"class="light-button drawing-control-button" >redo</button>
-					<button id="undo-button"class="light-button drawing-control-button" >undo</button>
+					<!-- <button id="redo-button"class="light-button drawing-control-button" >redo</button>
+					<button id="undo-button"class="light-button drawing-control-button" >undo</button> -->
 				</div>
 				<div class="middle-row">
 					<div id="drawing-canvas"></div>
@@ -29,9 +29,9 @@
 							<button id="move-down-button" class="code-button light-button">v</button>
 							<button id="add-above-button" class="code-button light-button">+ above</button>
 							<button id="add-below-button" class="code-button light-button">+ below</button>
-							<button id="remix-button" class="code-button light-button">remix</button>
+							<!-- <button id="remix-button" class="code-button light-button">remix</button> -->
 							<button id="shuffle-button" class="code-button light-button">shuffle</button>
-							<button id="group-button" class="code-button light-button">group</button>
+							<!-- <button id="group-button" class="code-button light-button">group</button> -->
 							<button id="delete-button" class="code-button light-button">delete</button>
 						</div>
 
@@ -47,4 +47,4 @@
 
 <!-- <script src="./libraries/joy.js"></script> -->
 <!-- <script src="./turtle.js"></script> -->
-<script type="module" src="./src/scripts.js"></script>
+<script type="module" src="./src/index.js"></script>
