+++ package.json
@@ -11,9 +11,10 @@
         "jest": "^28.1.3",
         "p5": "^1.4.2",
         "rangeslider.js": "^2.3.3",
-        "sortablejs": "^1.15.0"
+        "sortablejs": "^1.15.0",
+        "uuid": "^9.0.0"
     },
     "devDependencies": {
-        "snowpack": "^3.8.8"
+        "snowpack": "^3.7.1"
     }
 }
