+++ src/models/brush-params.js
@@ -1,0 +1,8 @@
+
+//TODO: pass around instead of "data" object to make sure everything is there
+//can also use for initial settings for each brush
+export class BrushParams {
+  constructor({color, size, alpha}) {
+
+  }
+}
