+++ src/models/brushstrokes.js
@@ -1,0 +1,202 @@
+import { v4 as uuidv4 } from 'uuid';
+
+export class BrushstrokeManager {
+  constructor() {
+    this.brushstrokes = [];
+  }
+
+  addBrushstroke(brushstroke) {
+    this.brushstrokes.push(brushstroke);
+    console.log("all brushstrokes: ", this.brushstrokes);
+    return brushstroke;
+  }
+
+  getBrushstrokeById(id) {
+    return this.brushstrokes.find(bs => bs.id === id);
+  }
+
+  getOnActByID(id) {  //return onAct function
+    return (settings) => {
+      const brushstroke = this.getBrushstrokeById(id);
+      if (brushstroke) {
+        brushstroke.renderFinal(settings);
+      }
+      else {
+        console.log('brushstroke not found');
+      }
+    }
+  }
+}
+
+export class Brushstroke {
+  constructor(effect, point, color, sketch) {
+    this.effect = effect;
+    console.log("effect", effect);
+    this.path = new Path(point.x, point.y);
+    this.color = color;
+    this.previewCanvas = sketch.getPreviewCanvas();
+    this.finalCanvas = sketch.getStaticCanvas();
+    this.pathComplete = false;
+
+    this.id = uuidv4();
+
+    console.log("new brushstroke with id " + this.id);
+  }
+
+  renderPreview(settings) {
+    console.log("rendering preview");
+    // could be a lower resolution render
+    // console.log("brushstroke points: ", this.path.getPoints());
+    this.effect.render(settings, this.path.getPoints(), this.previewCanvas);
+  }
+
+  renderFinal(settings) {
+    console.log("rendering final");
+    // TODO: if nothing has changed, use cached version
+    this.effect.render(settings, this.path.getPoints(), this.finalCanvas);
+  }
+
+  getPoints() {
+    // let points = [ this.path.getPoint(0) ];
+    // if(this.pathComplete) {
+    //   points = this.path.getPoints();
+    //   console.log("getting points!");
+    // }
+    // return points;
+    return this.path.getPoints();
+  }
+
+  getType() {
+    return this.effect.mouseActionType;
+  }
+
+  getID() {
+    return this.id;
+  }
+
+  // let brushAction = brushstrokeManager.getOnActByID(my.data.id);
+  //     brushAction(my.data);
+
+  makeJoyEvent(tag, userSettings) {
+    let settings = userSettings;
+    settings.id = this.id;
+
+    return [
+      tag,
+      this.effect.name,
+      settings
+    ]
+  }
+
+  addPoint(point) {
+    this.path.addPoint(point.x, point.y);
+  }
+
+  finalizePath() {
+    this.pathComplete = true;
+  }
+
+  makeAntPath(frogPath, stepSize) {
+    let antPath = [];
+    antPath.push({ position: frogPath.start });
+    let remainingDistance = stepSize;
+
+    for (let i = 1; i < frogPath.points.length; i++) {
+      let previousPoint = frogPath.points[i - 1].position;
+      let currentPoint = frogPath.points[i].position;
+
+      let vectorBetween = p5.Vector.sub(currentPoint, previousPoint);
+      let distanceBetween = vectorBetween.mag();
+      vectorBetween.normalize();
+
+      while (remainingDistance <= distanceBetween) {
+        let newPointPos = p5.Vector.add(
+          previousPoint,
+          vectorBetween.copy().mult(remainingDistance)
+        );
+        antPath.push({position: newPointPos});
+
+        distanceBetween -= remainingDistance;
+        remainingDistance = stepSize;
+        previousPoint = newPointPos;
+      }
+
+      remainingDistance -= distanceBetween;
+    }
+    return antPath;
+  }
+
+}
+
+export class Path {
+	constructor(x, y) {
+		this.points = [];
+		this.addPoint(x, y);
+	}
+	
+	addPoint(x, y) {
+		this.points.push({ x: x, y: y });
+	}
+
+  getPoints() {
+    return this.points;
+  }
+
+  getPoint(index) {
+    return this.points[index];
+  }
+
+  getLastPoint() {
+    return this.points[this.points.length - 1];
+  }
+
+	toString() {
+		let a = [];
+		this.points.forEach(p => {
+			a.push(Math.round(p.x));
+			a.push(Math.round(p.y));
+		});
+		return a.toString();
+	}
+	
+	renderPoints(s) {
+		s.push();
+		s.strokeWeight(5);
+		s.noFill();
+		s.stroke(0);
+		this.points.forEach(p => {
+			s.point(p.x, p.y);
+		});
+		s.pop();
+	}
+
+	renderLine(s, strokeColor, strokeWeight) {
+		s.push();
+		s.strokeWeight(strokeWeight);
+		s.stroke(strokeColor);
+		s.noFill();
+		s.beginShape();
+		this.points.forEach(p => {
+			s.vertex(p.x, p.y);
+		});
+		s.endShape();
+		s.pop();
+	}
+
+  // may not need this, placeholder from when paths were passed as strings
+  _stringToList(coordinates_string) {
+    const points = coordinates_string.split(",");
+    const result = [];
+  
+    for (let i = 0; i < points.length; i += 2) {
+      const x = parseInt(points[i].trim());
+      const y = parseInt(points[i + 1].trim());
+  
+      if (!isNaN(x) && !isNaN(y)) {
+        result.push({ x, y });
+      }
+    }
+  
+    return result;
+  }
+}
