+++ src/utils/utils.js
@@ -1,0 +1,6 @@
+export function reindex(list, key) {
+  return list.reduce((acc, item) => {
+      acc[item[key]] = item;
+      return acc;
+  }, {});
+}
