+++ test.js
@@ -1,0 +1,5 @@
+import sum from './src/index.js';
+
+test('adds 1 + 2 to equal 3', () => {
+    expect(sum(1, 2)).toBe(3);
+  });
